removed graphql and set up folder structure
