--- codegen.yml
@@ -1,23 +1,0 @@
-overwrite: true
-schema: "http://localhost:3000/api/graphql"
-documents: "./graphql/**/*.graphql"
-
-generates:
-  ./graphql/generated/index.tsx:
-    config:
-      reactApolloVersion: 3
-      withHooks: true
-    plugins:
-      - "typescript"
-      - "typescript-operations"
-      - "typescript-react-apollo"
-  # ./graphql/generated/validation.ts:
-  #   config:
-  #     typesPrefix: Z
-  #     withObjectType: true
-  #     importFrom: ./graphql/generated/index.tsx # path for generated ts code
-  #     schema: zod
-  #   plugins:
-  #     - "typescript"
-  #     - "typescript-operations"
-  #     - "typescript-validation-schema"
