+++ package.json
@@ -10,11 +10,14 @@
     "gen": "graphql-codegen --config codegen.yml"
   },
   "dependencies": {
-    "@apollo/client": "^3.6.9",
     "@formkit/auto-animate": "^1.0.0-beta.3",
     "@hookform/resolvers": "^2.9.7",
-    "apollo-server-core": "^3.10.1",
-    "apollo-server-micro": "^3.10.1",
+    "@prisma/client": "^4.8.0",
+    "@tanstack/react-query": "^4.20.4",
+    "@trpc/client": "^10.7.0",
+    "@trpc/next": "^10.7.0",
+    "@trpc/react-query": "^10.7.0",
+    "@trpc/server": "^10.7.0",
     "argon2": "^0.28.7",
     "axios": "^0.27.2",
     "connect-redis": "^6.1.3",
@@ -22,7 +25,6 @@
     "deepmerge": "^4.2.2",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
-    "graphql": "^15.8.0",
     "ioredis": "^5.2.2",
     "lodash": "^4.17.21",
     "micro": "^9.4.1",
@@ -38,16 +40,11 @@
     "react-query": "^3.39.2",
     "react-spinners": "^0.13.4",
     "reflect-metadata": "^0.1.13",
-    "type-graphql": "^1.1.1",
     "typeorm": "^0.3.7",
     "uuid": "^8.3.2",
     "zod": "^3.18.0"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^2.11.6",
-    "@graphql-codegen/typescript": "^2.7.3",
-    "@graphql-codegen/typescript-operations": "^2.5.3",
-    "@graphql-codegen/typescript-react-apollo": "^3.3.3",
     "@types/connect-redis": "^0.0.18",
     "@types/express": "^4.17.13",
     "@types/lodash": "^4.14.187",
@@ -59,7 +56,6 @@
     "autoprefixer": "^10.4.8",
     "eslint": "8.21.0",
     "eslint-config-next": "12.2.4",
-    "graphql-codegen-typescript-validation-schema": "^0.6.0",
     "postcss": "^8.4.16",
     "prettier": "^2.7.1",
     "prettier-plugin-tailwindcss": "^0.1.13",
