+++ src/pages/post/[postId].tsx
