+++ src/utils/loaders/createUserLoader.ts
@@ -1,6 +1,6 @@
 import DataLoader from "dataloader";
 import { In } from "typeorm";
-import { User } from "../../entities/User";
+import { User } from "../../server/entities/User";
 
 export const createUserLoader = () =>
   new DataLoader<string, User>(async (userIds) => {
